% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_apr}
\alias{find_apr}
\title{Find A-Phased Repeats}
\usage{
find_apr(dna_seq, minAPR = 4, maxAPR = 9, minATracts = 3)
}
\arguments{
\item{dna_seq}{DNA sequence string}

\item{minAPR}{Minimum length of A-tract (default: 4)}

\item{maxAPR}{Maximum length of A-tract (default: 9)}

\item{minATracts}{Minimum number of A-tracts (default: 3)}
}
\description{
Find A-Phased Repeats
}
