% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_dr}
\alias{find_dr}
\title{Find Direct Repeats}
\usage{
find_dr(dna_seq, mindir = 10, maxdir = 100, dspacer = 100)
}
\arguments{
\item{dna_seq}{DNA sequence string}

\item{mindir}{Minimum repeat size (default: 10)}

\item{maxdir}{Maximum repeat size (default: 100)}

\item{dspacer}{Maximum spacer size (default: 100)}
}
\description{
Find Direct Repeats
}
