% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_str}
\alias{find_str}
\title{Find Short Tandem Repeats}
\usage{
find_str(dna_seq, minSTR = 1, maxSTR = 6, minSTRlen = 10, minReps = 3)
}
\arguments{
\item{dna_seq}{DNA sequence string}

\item{minSTR}{Minimum unit size (default: 1)}

\item{maxSTR}{Maximum unit size (default: 6)}

\item{minSTRlen}{Minimum total length (default: 10)}

\item{minReps}{Minimum number of repeats (default: 3)}
}
\description{
Find Short Tandem Repeats
}
