% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_gq}
\alias{find_gq}
\title{Find G-quadruplexes}
\usage{
find_gq(dna_seq, minGQ = 3, maxGQspacer = 7)
}
\arguments{
\item{dna_seq}{DNA sequence string}

\item{minGQ}{Minimum G-run size (default: 3)}

\item{maxGQspacer}{Maximum loop size (default: 7)}
}
\description{
Find G-quadruplexes
}
