% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_ir}
\alias{find_ir}
\title{Find Inverted Repeats}
\usage{
find_ir(dna_seq, mincrf = 6, cspacer = 100, cut = 20, shortSpacer = 8)
}
\arguments{
\item{dna_seq}{DNA sequence string}

\item{mincrf}{Minimum repeat size (default: 6)}

\item{cspacer}{Maximum spacer size (default: 100)}

\item{cut}{Stem size threshold for short spacer check (default: 20)}

\item{shortSpacer}{Maximum spacer for short stems (default: 8)}
}
\description{
Find Inverted Repeats
}
