% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_mr}
\alias{find_mr}
\title{Find Mirror Repeats}
\usage{
find_mr(dna_seq, minmir = 10, mspacer = 100)
}
\arguments{
\item{dna_seq}{DNA sequence string}

\item{minmir}{Minimum repeat size (default: 10)}

\item{mspacer}{Maximum spacer size (default: 100)}
}
\description{
Find Mirror Repeats
}
