% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_zdna}
\alias{find_zdna}
\title{Find Z-DNA}
\usage{
find_zdna(dna_seq, minZ = 10)
}
\arguments{
\item{dna_seq}{DNA sequence string}

\item{minZ}{Minimum run size (default: 10)}
}
\description{
Find Z-DNA
}
