% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfa_analyze.R
\name{gfa_analyze}
\alias{gfa_analyze}
\title{Analyze all non-B DNA motifs in a sequence}
\usage{
gfa_analyze(dna_seq)
}
\arguments{
\item{dna_seq}{DNA sequence string}
}
\value{
A named list of data.frames for each motif type
}
\description{
Analyze all non-B DNA motifs in a sequence
}
