% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfaR-package.R
\docType{package}
\name{gfaR-package}
\alias{gfaR}
\alias{gfaR-package}
\title{gfaR: Genomic Feature Analyzer for Non-B DNA Motifs}
\description{
Refactored R package version of the Genomic Feature Analyzer (GFA) C-program for finding non-B DNA motifs in nucleic acid sequences.
}
\keyword{internal}
